+++ Cefalo.TechDaily.Service/Services/UserService.cs
@@ -12,7 +12,9 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Data.Common;
 using System.Runtime.ConstrainedExecution;
-
+using Cefalo.TechDaily.Service.Utils.Contracts;
+using Cefalo.TechDaily.Service.CustomExceptions;
+using FluentValidation;
 
 namespace Cefalo.TechDaily.Service.Services
 {
@@ -21,20 +23,23 @@ namespace Cefalo.TechDaily.Service.Services
         private readonly IUserRepository _userRepository;
         private readonly IMapper _mapper;
         private readonly IPasswordHandler _passwordHandler;
-        public UserService(IUserRepository userRepository, IMapper mapper, IPasswordHandler passwordHandler)
+        private readonly IJwtTokenHandler _jwtTokenHandler;
+        public UserService(IUserRepository userRepository, IMapper mapper, IPasswordHandler passwordHandler, IJwtTokenHandler jwtTokenHandler)
         {
             _userRepository = userRepository;
             _mapper = mapper;
             _passwordHandler = passwordHandler;
+            _jwtTokenHandler = jwtTokenHandler;
         }
         public async Task<List<UserDto>> GetUsers()
         {
             var users = await _userRepository.GetUsers();
             return users.Select(user => _mapper.Map<UserDto>(user)).ToList();
         }
-        public async Task<UserDto?> GetUserByUsername(string Username)
+        public async Task<UserDto> GetUserByUsername(string Username)
         {
             var user = await _userRepository.GetUserByUsername(Username);
+            if (user == null) throw new NotFoundException("User not found");
             return _mapper.Map<UserDto>(user);
         }
         public async Task<UserDto> PostUser(SignupDto request)
@@ -47,13 +52,21 @@ namespace Cefalo.TechDaily.Service.Services
             user.PasswordHash = passwordHash;
             user.PasswordSalt = passwordSalt;
             var newUser = await _userRepository.PostUser(user);
+            if (newUser == null) throw new BadRequestException("Cant add user");
             var userDto = _mapper.Map<UserDto>(newUser);
             return userDto;
         }
 
-        public async Task<UserDto?> UpdateUser(string Username, UpdateUserDto updateUserDto)
+        public async Task<UserDto> UpdateUser(string Username, UpdateUserDto updateUserDto)
         {
-
+            var currentUser = await _userRepository.GetUserByUsername(Username);
+            if (currentUser == null) throw new NotFoundException("User not found");
+            var loggedInUser = _jwtTokenHandler.GetLoggedinUsername();
+            if (loggedInUser != Username) throw new UnauthorizedException("You are not authorized to update this user's information");
+            var tokenCreationTimeString = _jwtTokenHandler.GetTokenCreationTime();
+            if (tokenCreationTimeString == null) throw new UnauthorizedException("Please login again");
+            DateTime tokenCreationTime = Convert.ToDateTime(tokenCreationTimeString); 
+            if(DateTime.Compare(tokenCreationTime,currentUser.PasswordModifiedAt)<0) throw new UnauthorizedException("Please login again");
             User user = _mapper.Map<User>(updateUserDto);
             if (updateUserDto.Password != null)
             {
@@ -64,13 +77,21 @@ namespace Cefalo.TechDaily.Service.Services
             }
             user.UpdatedAt = DateTime.UtcNow;
             var newUser = await _userRepository.UpdateUser(Username, user);
-            if (newUser == null) return null;
             var userDto = _mapper.Map<UserDto>(newUser);
             return userDto;
         }
-        public Task<Boolean> DeleteUser(string Username)
+        public async Task<Boolean> DeleteUser(string Username)
         {
-            return _userRepository.DeleteUser(Username);
+            var currentUser = await _userRepository.GetUserByUsername(Username);
+            if (currentUser == null) throw new NotFoundException("User not found");
+            var loggedInUser = _jwtTokenHandler.GetLoggedinUsername();
+            if (loggedInUser != Username) throw new UnauthorizedException("You are not authorized to delete this user");
+            var tokenCreationTimeString = _jwtTokenHandler.GetTokenCreationTime();
+            if (tokenCreationTimeString == null) throw new UnauthorizedException("Please login again");
+            DateTime tokenCreationTime = Convert.ToDateTime(tokenCreationTimeString);
+            if (DateTime.Compare(tokenCreationTime, currentUser.PasswordModifiedAt) < 0) throw new UnauthorizedException("Please login again");
+
+            return await _userRepository.DeleteUser(Username);
         }
